pyproject.toml
[project]
name = "mcp-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.5.1",
    "openai>=1.82.0",
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.45.1",
]
